Converted App.js to App.tsx
